refactor: reimplements `walletconnect` v1 & v2 providers to work simultaneously based on init opts
